Message

+++ arya_mathloop.c
@@ -25,11 +25,11 @@ while((getchar())!= '\n');
 
 //Operating the switch appropriately
 switch (ch) 
-{case '+': f = a + b; printf("f = %d\n", f); break;
-case '-': f = a - b; printf("f = %d\n", f); break; 
-case '*': f = a * b; printf("f = %d\n", f); break; 
+{case '+': f = (double)a + b; printf("f = %d\n", f); break;
+case '-': f = (double)a - b; printf("f = %d\n", f); break; 
+case '*': f = (double)a * b; printf("f = %d\n", f); break; 
 case '/': f = ((double)a / b); printf("f = %f\n", f); break;
-case '%': f = a%b;  printf("f = %d\n", f); break;
+case '%': f = (double)a%b;  printf("f = %d\n", f); break;
 case 'q': break;
 default: printf("invalid operator\n");  }
